requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
entrypoints==0.3
h11==0.12.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
outcome==1.1.0
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py==1.10.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1

pyzmq==22.2.1
requests==2.26.0
selenium==4.0.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.5.1
webencodings==0.5.1
wsproto==1.0.0
